update: adjust SearchModal styled
