+++ src/components/Header/styled.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { pagePadding } from "../../constants/style/commonStyled";
+import { cardWrapper, pagePadding } from "../../constants/style/commonStyled";
 
 export const Wrapper = styled.div`
   display: flex;
@@ -9,4 +9,9 @@ export const Wrapper = styled.div`
   width: 100%;
   padding: 24px;
   ${pagePadding}
+`;
+
+export const SearchButton = styled.button`
+  padding: 0 8px;
+  ${cardWrapper}
 `;
